chore(modules-sdk): Move Query and cleanup
